requirements.txt
ago==0.0.95
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
datasets>=2.14.5
dotmap==1.3.30
elastic-transport==8.4.0
elasticsearch==8.8.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
hjson==3.1.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.2.0
langdetect==1.0.9
lxml==4.9.2
news-please @ git+https://github.com/lucadiliello/news-please.git
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.1
parsel==1.8.1
Pillow==9.5.0
plac==1.3.5
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyMySQL==1.0.3
pyOpenSSL==23.2.0
python-dateutil==2.8.2
PyYAML==6.0
queuelib==1.6.2
readability-lxml==0.8.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
s3transfer==0.6.1
Scrapy==2.9.0
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.6.3
urllib3==1.26.16
w3lib==2.1.1
warcio==1.7.4
zope.interface==6.0
